% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_select.R
\name{color_select}
\alias{color_select}
\title{Create a widget that selects which parameter to color markers by}
\usage{
color_select()
}
\value{
A select input widget with the options to color by different variables
}
\description{
Create a widget that selects which parameter to color markers by
}
