% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_filter.R
\name{conditional_filter}
\alias{conditional_filter}
\title{Return a result conditionally}
\usage{
conditional_filter(condition, success)
}
\arguments{
\item{condition}{A conditional statement}

\item{success}{The result if the conditional statement returns TRUE}
}
\value{
The success parameter if the condition is met, or TRUE if not
}
\description{
\code{conditional_filter()} returns a specified result if the condition is met, or TRUE otherwise
}
\details{
This function is used to reactively filter a dataframe with \code{dplyr::filter()}.
If the condition is met, the filtering condition specified in \code{success} will be returned,
filtering the dataframe. If the condition is not met, it will return TRUE.
Because \code{filter()} accepts any row that returns TRUE, this returns the unfiltered dataframe.
}
\examples{
# Condition returns TRUE, filters successfully
x <- 5
df <- dplyr::filter(accidents, conditional_filter(x == 5, day.night == "night"))

# Condition returns FALSE, filter() sees TRUE, so does not filter dataframe
x <- 3
df <- dplyr::filter(accidents, conditional_filter(x == 5, day.night == "night"))

}
