% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daynight_select.R
\name{daynight_select}
\alias{daynight_select}
\title{Create a widget that allows you to select whether to filter data by day/night}
\usage{
daynight_select()
}
\value{
A select input widget that filters accidents by whether it occurred in the day or night
}
\description{
Create a widget that allows you to select whether to filter data by day/night
}
