% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_if.R
\name{filter_if}
\alias{filter_if}
\title{Filter only if a condition is met}
\usage{
filter_if(condition, success)
}
\arguments{
\item{condition}{A conditional statement}

\item{success}{The result if the conditional statement returns TRUE}
}
\value{
The success parameter if the condition is met, or TRUE if not
}
\description{
Filter only if a condition is met
}
\examples{
# Condition returns TRUE, filters successfully
x <- 5
df <- accidents \%>\%
dplyr::filter(filter_if(x == 5, day.night == "night"))

# Condition returns FALSE, filter() sees TRUE, so does not filter dataframe
x <- 3
df <- accidents \%>\%
dplyr::filter(filter_if(x == 5, day.night == "night"))

}
