% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Choose to plot histogram or bar plot}
\usage{
histogram_plot(x, y)
}
\arguments{
\item{x}{A variable from \code{accidents} to plot}

\item{y}{The number of bins}
}
\value{
A histogram plot if continuous or a bar plot if categorical
}
\description{
\code{histogram_plot()} plots a histogram if a continuous variable from \code{accidents}
is chosen or a bar plot if a categorical variable is chosen.
}
\details{
Creates a bar plot if a categorical variable is chosen to prevent issues
with plotting non-continuous data as a histogram.
}
