% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hour_select.R
\name{hour_select}
\alias{hour_select}
\title{Create a widget that selects whether to filter by hour}
\usage{
hour_select()
}
\value{
A selector widget that filters accidents by hour they occurred
}
\description{
Create a widget that selects whether to filter by hour
}
