% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_title.R
\name{legend_title}
\alias{legend_title}
\title{Select a legend title}
\usage{
legend_title(c)
}
\arguments{
\item{c}{An input that determines which title is returned}
}
\value{
A character string representing a legend title
}
\description{
\code{legend_title()} selects a title from a list depending on the input.
}
\details{
This function is to return a legend title that corresponds with the variable
chosen to color data points by.
}
