% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_select.R
\name{month_select}
\alias{month_select}
\title{Create a widget that select whether to filter by month}
\usage{
month_select()
}
\value{
A selector widget that filters accidents by month they occurred
}
\description{
Create a widget that select whether to filter by month
}
