% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_input.R
\name{select_input}
\alias{select_input}
\title{Create selector for a shiny app}
\usage{
select_input(inputId, label, choices, type)
}
\arguments{
\item{inputId}{A shiny input ID}

\item{label}{A label for the input}

\item{choices}{A vector of choices for the input}

\item{type}{The type of input, select for selectInput, multiple for selectInput multiple = TRUE, and checkbox for checkboxGroupInput}
}
\value{
A shiny widget that allows user to select from a list of choices
}
\description{
\code{select_input()} creates a specified shiny UI input.
}
\details{
This function allows multiple different types of shiny UI input widgets to be
created depending on the value of \code{type}
}
\examples{
select_input("color", "Color By:", c("None", "Temp", "Day/Night"), "select")
}
