% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_select.R
\name{weather_select}
\alias{weather_select}
\title{Create a widget that selects whether to filter data by weather type}
\usage{
weather_select()
}
\value{
A selector widget that filters accidents based on the category of weather
}
\description{
Create a widget that selects whether to filter data by weather type
}
