% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pal.R
\name{color_pal}
\alias{color_pal}
\title{Select a color pallette}
\usage{
color_pal(a)
}
\arguments{
\item{a}{A character string}
}
\value{
A color pallette assigned to the values of a column in the \code{accidents} dataset
}
\description{
\code{color_pal()} returns a color pallette with values assigned to the values
of a column in the \code{accidents} dataset.
}
\details{
This function returns a color pallette based on the input in the \code{a} argument.
It is designed to color markers for a leaflet map in a shiny app based on user selection.
}
