% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Switch between histogram or bar plot based on type of data}
\usage{
histogram_plot(x, y)
}
\arguments{
\item{x}{A variable to plot}

\item{y}{The number of bins}
}
\value{
A histogram plot if continuous or a bar plot if categorical
}
\description{
Switch between histogram or bar plot based on type of data
}
