% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_input.R
\name{select_input}
\alias{select_input}
\title{Create input for a shiny app}
\usage{
select_input(inputId, label, x, type)
}
\arguments{
\item{inputId}{A shiny input ID}

\item{label}{A label for the input}

\item{x}{A vector of choices for the input}

\item{type}{The type of input, select for selectInput, multiple for selectInput multiple = TRUE, and checkbox for checkboxGroupInput}
}
\value{
A shiny widget that allows user to select from a list of choices
}
\description{
Create input for a shiny app
}
\examples{
select_input("color", "Color By:", c("None", "Temp", "Day/Night"), "select")
}
