% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_options.R
\name{select_options}
\alias{select_options}
\title{Create a character vector of column values}
\usage{
select_options(df, col, all = FALSE, none = FALSE)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The name of a column from the dataframe}

\item{all}{Default is FALSE, when TRUE will add "All" to the vector}

\item{none}{Default is FALSE, when TRUE will add "None" to the vector}
}
\value{
A character vector
}
\description{
\code{select_options()} takes a column of a dataframe and creates a character vector
of all unique values within that column.
}
\details{
Used to create a list of options from the unique values within a column.
}
